+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/DomainTest.java
@@ -560,7 +560,7 @@ public class DomainTest extends AbstractModelTest {
 		Assert.assertEquals(s2, m.getConfigValue());
 		Assert.assertEquals(d1, m.getCreated());
 		Assert.assertEquals(d2, m.getModified());
-		Assert.assertEquals(s3, m.getUserId());
+		Assert.assertEquals(s3, m.getUserId());		
 	}
 
 	@Test
@@ -572,6 +572,7 @@ public class DomainTest extends AbstractModelTest {
 		m.setCreated(d1);
 		m.setModified(d2);
 		m.setUserId(s3);
+		
 		checkMLPSiteConfig(m);
 		m = new MLPSiteConfig(m);
 		checkMLPSiteConfig(m);