+++ docs/release-notes.rst
@@ -27,6 +27,7 @@ Version 1.15.0, 27 March 2018
 -----------------------------
 
 * Move acccess type code attribute from Solution to Solution Revision entity (ACUMOS-196)
+* Add fields to site_config table entry: validation (ACUMOS-486)
 * Requires database schema version 1.15
 
 Version 1.14.2, 15 March 2018